Chore: Move core rangeutil tests into grafana-data
